insert committee when handling checkpoint from consensus (#12324)
## Description 

In https://github.com/MystenLabs/sui/pull/12212, we refactored
`handle_checkpoint_from_consensus`. The PR missed one thing: when a last
checkpoint of an epoch is certified in consensus and sent to state sync,
we do not update the committee store. This is problematic because state
sync does not reconfig. It means if a validator is slow in
reconfiguration (where committee store would also be updated) and gets
checkpoints in the next epoch from peers, it will panic in that it
couldn't find the committee. This PR fixes it.


## Test Plan 

```
for f in `seq 1 20`; do MSIM_TEST_NUM=5 MSIM_TEST_SEED=98$f RUST_LOG=off simt test_simulated_load_reconfig_restarts --no-capture > log-$f 2>&1 &; done
```

```
grep FAIL log-*
```
---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes
Insert committee when handling checkpoint from consensus.