Add storage format enum to file metadata in archive manifest (#12325)
## Description 

Right now we always assume the underlying storage format is blob. This
allows us to migrate to with other storage formats in the future without
needing to migrate older data to newer format. We would be able to just
add a new enum value and start writing newer checkpoints with newer
storage format.

## Test Plan 

Existing tests