[ts-sdk] Use protocol config in transaction builder (#12081)
## Description 

- Adds `getProtocolConfig()`
- Reworks the hard-coded limits to instead use the protocol config. As a
performance optimization, a `protocolConfig` option can be used instead
of having it fetch whenever it builds the transaction. Additionally, you
can hard-code `limits` if you wish, to entirely avoid working with
protocol config (not recommended, but can be used to override protocol
config if you wanted to for whatever reason).
- I had to rework the limit for pure inputs so that it happens as
validation instead of during serialization to bytes. This means that
pure inputs are completely unlimited (can grow to infinity).
- Some things that previously errored during setting properties will now
error during transaction construction to bytes (for example, setting
payment coins).

## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes

---------

Co-authored-by: Michael Hayes <michael.hayes@mystenlabs.com>