[narwhal] introduce the LeaderSwapTable (#12465)
## Description 

As part of the leader election schedule change epic
https://www.notion.so/mystenlabs/Leader-election-schedule-31e9abd06477436b9a38fd3282879093?pvs=4
, the first step is to build a "swap" table based on the calculated
reputation scores. As a reminder we do calculate the leader reputation
scores every K commit rounds. Our goal is based on those scores to come
up with a set of good & bad nodes. Then for next K commit rounds we do
swap the occurrences of leaders of the bad nodes with the ones of the
good nodes.

This PR is the first step. As part of the upcoming work we'll need to 
- [ ] calculate the swap table on every K committed subdags & wire into
the leader election algorithm
- [ ] restore correct swap table after crash/recovery
- [ ] modify the commit path so we repeat the leader election when we
commit recursively
- [ ] modify the proposer to support the new leader election
capabilities
- [ ] ensure whole feature is gated behind a protocol config feature
flag - and probably a config switch as we might need to have it disabled
for longer than a release cycle.
- [ ] add testing

## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes