Perform full checkpoint pruning at epoch boundary in simtests (#12529)
## Description 

This is just to be extra sure that we are testing the worst case tx/fx
pruning in order to maximize hitting edge cases. At epoch boundary,
after the epoch store has been swapped, we prune all eligible
transactions and effects (which should be all existing transactions and
effects at that point), and ensure that we can continue. This more
strongly confirms that the protocol does not depend on prune-eligible
transactions and effects at any point.

## Test Plan 

All reconfig simtests will exercise this

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes