move post processing before critical commits (#12669)
## Description 

1. move post processing before critical commits. This makes sure the
post processing wouldn't run out of causal order.
2. change some `continue`s to panic so bugs will scream loudly.
3. pass `inner_temp_store.written` to post processing code such that
latest objects become accessible without looking into DB (which is not
ready at that point)
4. deprecates `timestamps` indexer column which is not meaningful and in
use.

## Test Plan 

Apply this to 1.4.0 and resync from genesis. 

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes
This change is fullnode only and does not impact validator. If a
fullnode serves RPC requests based on secondary indices, they may want
to re-sync from genesis with this binary to calculate all the data
right. There will also be a snapshot ready for consumption.